Backend/requirements.txt
django==4.2.8
djangorestframework==3.14.0
mysqlclient==2.2.8
importlib-metadata
pandas==2.2.3
openpyxl==3.1.2
